requirements.txt
pyrogram==2.0.106
tgcrypto
anthropic>=0.40.0
python-dotenv>=1.0.0
